requirements.txt
Django==3.2.14
python-dotenv==1.0.0
requests==2.31.0
django-crispy-forms==2.0
crispy-tailwind==0.5.0
pymongo==4.6.1
djongo==1.3.6
discord.py==2.3.2
python-social-auth==0.3.6
social-auth-app-django==5.4.0
Pillow==10.1.0
django-tailwind==3.6.0
